use cobra for cmd flags
